data-worker/requirements.txt
redis==5.0.1
Faker==23.2.1
python-dotenv==1.0.1